Fix caret move unavailable during edit and error message display cut off on large list
Caret moving with the arrow key or finger on smartphone was impossible because draggable=true on parent div element breaks this. Workaround is temporary set draggable=false on parent div while edit is active.

When a large list was scrolled to bottom and a modal message is shown user could not see the text. Workaround is remember current scroll offsets, scroll to top, show modal, and restore scroll offsets when modal closed.